+++ Modules/Account/moduleSrc/ORM/ActiveSessions.cpp
@@ -58,7 +58,7 @@ ActiveSessions::ActiveSessions() :
            tblActiveSessions::Name,
             { ///<ColName                              Type                    Validation                   Default    UpBy   Sort  Filter Self  Virt   PK
               {tblActiveSessions::ssnKey,              S(TAPI::MD5_t),        QFV,                         ORM_PRIMARY_KEY},
-              {tblActiveSessions::ssn_usrID,           S(quint32),            QFV.integer().minValue(1),   QRequired,   UPNone},
+              {tblActiveSessions::ssn_usrID,           S(quint64),            QFV.integer().minValue(1),   QRequired,   UPNone},
               {tblActiveSessions::ssnIP,               S(quint32),            QFV.integer().minValue(1),   QRequired,   UPNone},
               {tblActiveSessions::ssnIPReadable,       S(QString),            QFV.allwaysInvalid(),        QInvalid,   UPNone,false,false},
               {tblActiveSessions::ssnCreationDateTime, ORM_CREATED_ON},
